Debugger v0.2
    Debugger v0.2

    - Windows: Disasm, copper, blitter, custom regs, scr_graph, console;
    - Shortcuts, Actions, Log, Console, messages;
    - Debug trace: Step-In, Step Copper, Skip N scanlines;
    - Separate thread for UAE
    - Update imgui docking to latest ver 1.91.3
    - and many, many ..

Fix creation of UAE window and process events in main thread

- Due to bug on MacOS: NSWindow should only be instantiated on the main thread!
- A known bug is that shortcuts are now global instead of separate for each window.

+++ src/adf.cpp
@@ -106,7 +106,7 @@ void Adf::create_for_exefile(const char* name) {
         exit(1);
     }
 
-    adfFileWrite(file, size, buffer);
+    adfFileWrite(file, (uint32_t)size, buffer);
     adfFileClose(file);
 
     // Create the S directory
@@ -126,7 +126,7 @@ void Adf::create_for_exefile(const char* name) {
     // Create the Startup-Sequence file and write commands to it
     file = adfFileOpen(volume, "Startup-Sequence", ADF_FILE_MODE_WRITE);
     if (file) {
-        adfFileWrite(file, strlen(name), (unsigned char*)name);
+        adfFileWrite(file, (uint32_t)strlen(name), (unsigned char*)name);
         adfFileClose(file);
     } else {
         printf("Error creating Startup-Sequence file\n");
